requirements.txt
gradio == 3.27.0
openai == 0.27.4
python-dotenv == 1.0.0
langchain == 0.0.148
#chromadb == 0.3.21
tiktoken == 0.3.3
pypdf == 3.8.1
pymupdf == 1.22.2
pinecone-client == 2.2.2